backend/requirements.txt
fastapi[all]
mysql-connector-python
python-dotenv